Allow internal modules to be importable by external consumers
Note here that this is a simple rename from `internal/` --> `pkg/`.
Longer-term, much of this code should probably live within the top-level
of this module, but given some of the internal packages have unclear
names, I'd prefer to not take this on at the moment and allow the tool
maintainers to have some discussion about future module layout.

Signed-off-by: Stephen Augustus <foo@auggie.dev>
Co-authored-by: Dan Lorenc <dlorenc@google.com>