Add runner package skeleton
This commit adds the skeleton of a mockable runner package.

The package has a generator object that executes all the SBOM generation
logic. It relies on two underlying objects: the implementation and the
document format handler.

- The implementation handles all the generator main functions.
- All the format-specific code lies in the format handler which can be
swapped out for each format supported by the generator.

This is a skeleton only, the logic needs to be implemented still.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>